render_spriteで四角の位置計算途中。近寄ったときの数値がおかしい

+++ includes/structs/struct_sprite.h
@@ -6,7 +6,7 @@
 /*   By: rnakai <rnakai@student.42tokyo.jp>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/24 15:38:55 by rnakai            #+#    #+#             */
-/*   Updated: 2020/11/26 10:38:37 by rnakai           ###   ########.fr       */
+/*   Updated: 2020/11/26 12:08:12 by rnakai           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ typedef struct
 	int			x; //spriteのx座標。
 	int			y;
 	float		distance; //spriteとプレイヤーの距離
-	int			center_x_to_render; //スプライトの描画の中心x座標。ここを起点に描画
+	float		center_x_to_render; //スプライトの描画の中心x座標。ここを起点に描画
 	float		left_len; //中央から見て左側の長さ
 	float		right_len; //右側の長さ
 }		t_sprite;
